tested with local ollama container, working
