+++ src/services/llms/Ollama.py
@@ -1,5 +1,5 @@
 # -------------------------------------------------------------------------------------------------------------
-# File: llm.py
+# File: Ollama.py
 # Project: Open Source Institute-Cognitive System of Machine Intelligent Computing (OpenSI-CoSMIC)
 # Contributors:
 #     Muntasir Adnan <adnan.adnan@canberra.edu.au>
@@ -26,22 +26,20 @@ sys.path.append(f"{os.path.dirname(os.path.abspath(__file__))}/../../..")
 
 import os
 import sys
-import json
 import requests
 import subprocess
-from typing import Optional, List, Dict, Any
+from typing import Optional, List, Dict, Any, Union
 
 from src.services.llms.llm import LLMBase
 
 class Ollama(LLMBase):
-    def __init__(
-        self,
-        llm_name: str,
-        container_name: str = "ollama",
-        local_port: int = 11434,
-        **kwargs
-    ):
-        """For Ollama supported LLMs using the official Ollama container.
+    def __init__(self,
+                 llm_name: str = "llama3.2",
+                 container_name: str = "ollama",
+                 local_port: int = 11434,
+                 **kwargs) -> None:
+        """
+        LLMs using the official Ollama container.
 
         Args:
             llm_name (str, optional): Ollama supported LLMs available at https://ollama.com/library.
@@ -50,17 +48,13 @@ class Ollama(LLMBase):
         """
         # Extract the model name from "ollama:model_name" format if needed
         model_name = llm_name.replace("ollama:", "")
-        
-        # Call parent constructor with the clean model name
         super().__init__(llm_name=model_name, **kwargs)
-        
-        # Container configuration
         self.container_name = container_name
         self.local_port = local_port
         # self.api_url = f"http://{self.container_name}:{self.local_port}/api"
-        self.api_url = f"http://localhost:{self.local_port}/api/generate" # use chat for chat mode
+        # https://github.com/ollama/ollama/blob/main/docs/api.md
+        self.api_url = f"http://localhost:{self.local_port}/api/generate"
         
-        # Check if model is available, pull if needed
         self._check_model_availability()
 
 
@@ -70,6 +64,9 @@ class Ollama(LLMBase):
 
         Args:
             model_list (str): The output of the model list command.
+            
+        Returns:
+            List[str]: A list of model names.
         """
         models = []
         # Skip header line and get just the names
@@ -84,9 +81,13 @@ class Ollama(LLMBase):
             
         return models
     
+    
     def _get_model_list(self) -> List[str]:
         """
-        Get the list of models available on the ollama server.
+        Get the list of models available on the ollama container.
+        
+        Returns:
+            List[str]: A list of model names available on the ollama server.
         """
         shell_command = ["docker", "exec", self.container_name, "ollama", "list"]
         
@@ -95,12 +96,14 @@ class Ollama(LLMBase):
                                     capture_output=True)
         
         if model_list.returncode != 0:
-            print(f"Error executing docker command: {model_list.stderr}")
+            print("Error fetching the model list.")
             return []
             
+        # Process
         return self._process_model_list(model_list.stdout)
     
-    def _pull_model(self):
+    
+    def _pull_model(self) -> None:
         """
         Pull the model from the server using secure subprocess execution
         """
@@ -118,7 +121,8 @@ class Ollama(LLMBase):
         
         print(f"Model {self.llm_name} pulled successfully.")
     
-    def _check_model_availability(self):
+    
+    def _check_model_availability(self) -> None:
         """
         Check if the model is available on the ollama server.
         """
@@ -129,12 +133,14 @@ class Ollama(LLMBase):
         else:
             self._pull_model()
     
+    
     def _format_messages(self, messages: List[Dict[str, str]]) -> str:
         """
-        Format messages into a single prompt string.
+        Format messages into a single prompt string for /generate.
         
         Args:
-            messages: List of message dictionaries with 'role' and 'content' keys
+            messages: List of message dictionaries with 'role' and 'content' keys.
+            Only 'system' and 'user' roles are considered.
             
         Returns:
             A formatted string combining all messages
@@ -151,7 +157,7 @@ class Ollama(LLMBase):
             
             # Format based on role
             if role == 'system':
-                prompt += f"System:\n{content.strip()}\n\n"
+                prompt += f"System:\n{content.strip()}\n"
             elif role == 'user':
                 prompt += f"User:\n{content.strip()}\n\n"
         
@@ -160,7 +166,8 @@ class Ollama(LLMBase):
         
         return prompt
     
-    def _send_chat_request(self, messages: str) -> Dict[str, Any]:
+    
+    def _send_chat_request(self, prompt: str) -> Dict[str, Any]:
         """
         Send a chat request to the Ollama API.
         
@@ -170,50 +177,50 @@ class Ollama(LLMBase):
         Returns:
             The API response as a dictionary
         """
+        # 'messages' for /chat endpoint
         payload = {
             "model": self.llm_name,
-            "messages": messages,
-            # "prompt": "what is the capital of France?",
+            "prompt": prompt,
             "stream": False
         }
         
         try:
             response = requests.post(f"{self.api_url}", json=payload)
-            response.raise_for_status()  # Raise exception for HTTP errors
+            # Raise exception for HTTP errors
+            response.raise_for_status()
             return response.json()
         except requests.exceptions.RequestException as e:
             print(f"Error communicating with Ollama API: {e}")
             raise RuntimeError(f"Failed to get response from Ollama: {e}")
+        
     
-    def __call__(
-        self, 
-        question: str, 
-        context: dict = {}
-    ):
+    def __call__(self,
+                 question: str,
+                 context: Union[str, Dict[str, Any]] = {}) -> tuple[str, str]:
         """
         Process the question and generate a response using Ollama.
         
         Args:
             question (str): User question in string.
-            context (dict, optional): Context retrieved externally if applicable.
+            context (Union[str, Dict[str, Any]], optional): Context for the question. Defaults to {}.
             
         Returns:
             Tuple of (response, raw_response)
         """
         # Generate user prompt with question and context
-        user_prompt = self.user_prompter(question, context=context)
+        user_prompt: str = self.user_prompter(question, context=context)
         
         # Combine system prompt with user prompt
-        system_prompt = self.system_prompter(user_prompt, context=context)
+        combined_prompt: list[dict] = self.system_prompter(user_prompt, context=context)
         
-        # Format messages for Ollama
-        messages = self._format_messages(system_prompt)
-        
+        # Convert prompt to string
+        prompt: str = self._format_messages(combined_prompt)
         # Send request to Ollama
-        response_data = self._send_chat_request(system_prompt[0])
+        response_data: dict = self._send_chat_request(prompt)
         
         # Extract the response content
-        raw_response = response_data.get('message', {}).get('content', '')
+        # raw_response = response_data.get('message', {}).get('content', '')
+        raw_response = response_data.get('response', '')
         
         # Apply truncation if enabled
         response = self.truncate_response(raw_response) if self.is_truncate_response else raw_response
@@ -228,7 +235,7 @@ class Ollama(LLMBase):
     
 
 if __name__ == "__main__":
-    # Example usage
-    ollama = Ollama(llm_name="ollama:llama3.2", container_name="condescending_gould", local_port=11433)
-    response = ollama("What is the capital of France?")
-    print(response)  # Should print the response from the model
+    ollama = Ollama(local_port=11433)
+    response, raw_response = ollama.__call__("What is the capital of France?")
+    print(response)
+    print(raw_response)
